# api/docs/swagger.yaml
definitions:
...
  http.Response:
    properties:
      data: {}
      description:
        type: string
      status:
        type: string
    type: object
...
  users_service.CreateUser:
    properties:
      first_name:
        type: string
      last_name:
        type: string
      telegram_id:
        type: string
      username:
        type: string
    type: object
  users_service.GetListUserBuyTransactionResponse:
    properties:
      count:
        type: integer
      user_transaction:
        items:
          $ref: '#/definitions/users_service.UserBuyTransaction'
        type: array
    type: object
  users_service.GetListUserResponse:
    properties:
      count:
        type: integer
      users:
        items:
          $ref: '#/definitions/users_service.User'
        type: array
    type: object
  users_service.GetListUserSellTransactionResponse:
    properties:
      count:
        type: integer
      user_transaction:
        items:
          $ref: '#/definitions/users_service.UserSellTransaction'
        type: array
    type: object
...
  users_service.UpdateUser:
    properties:
      id:
        type: string
      status:
        type: string
    type: object
  users_service.User:
    properties:
      created_at:
        type: string
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      status:
        type: string
      telegram_id:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  users_service.UserBuyTransaction:
    properties:
      CheckImg:
        type: string
      all_price:
        type: string
      coin:
        type: string
      coin_amount:
        type: string
      coin_price:
        type: string
      created_at:
        type: string
      id:
        type: string
      message:
        type: string
      status:
        type: string
      user:
        type: string
      user_address:
        type: string
    type: object
  users_service.UserSellTransaction:
    properties:
      CheckImg:
        type: string
      all_price:
        type: string
      card_holder_name:
        type: string
      card_number:
        type: string
      coin_amount:
        type: string
      coin_id:
        type: string
      coin_price:
        type: string
      created_at:
        type: string
      id:
        type: string
      message:
        type: string
      status:
        type: string
      user_id:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /user:
    get:
      consumes:
      - application/json
      description: Get Users List
      operationId: get_user_list
      parameters:
      - description: offset
        in: query
        name: offset
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetAllUserResponseBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/users_service.GetListUserResponse'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get Users List
      tags:
      - User
    post:
      consumes:
      - application/json
      description: Create User
      operationId: create_user
      parameters:
      - description: CreateUserBody
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/users_service.CreateUser'
      produces:
      - application/json
      responses:
        "200":
          description: GetUserBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/users_service.User'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Create User
      tags:
      - User
  /user/{id}:
    get:
      consumes:
      - application/json
      description: Get User  By ID
      operationId: get_user_by_id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: UserBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/users_service.User'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get User  By ID
      tags:
      - User
    put:
      consumes:
      - application/json
      description: Update User
      operationId: update_user
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: UpdateUser
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/users_service.UpdateUser'
      produces:
      - application/json
      responses:
        "200":
          description: User data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/users_service.User'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Update User
      tags:
      - User
  /user/buy:
    get:
      consumes:
      - application/json
      description: Get Transactions List
      operationId: get_transaction_by_list
      parameters:
      - description: offset
        in: query
        name: offset
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      - description: status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetAllCoinResponseBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/users_service.GetListUserBuyTransactionResponse'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get Transactions List
      tags:
      - User
    post:
      consumes:
      - application/json
      description: UserBuy
      operationId: User_buy
      parameters:
      - description: Upload file
        in: formData
        name: file
        required: true
        type: file
      - description: User ID
        in: formData
        name: userId
        required: true
        type: string
      - description: Coin ID
        in: formData
        name: coinId
        required: true
        type: string
      - description: Address
        in: formData
        name: address
        required: true
        type: string
      - description: CoinAmount
        in: formData
        name: coin_amount
        required: true
        type: string
      - description: Message
        in: formData
        name: message
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetUserBuyBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: UserBuy
      tags:
      - User
...
  /user/sell:
    get:
      consumes:
      - application/json
      description: Get Transactions List
      operationId: get_transaction_sell_list
      parameters:
      - description: offset
        in: query
        name: offset
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      - description: status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetAllCoinResponseBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/users_service.GetListUserSellTransactionResponse'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get Transactions List
      tags:
      - User
    post:
      consumes:
      - application/json
      description: UserSell
      operationId: User_sell
      parameters:
      - description: Upload file
        in: formData
        name: file
        required: true
        type: file
      - description: User ID
        in: formData
        name: userId
        required: true
        type: string
      - description: Coin ID
        in: formData
        name: coinId
        required: true
        type: string
      - description: CoinAmount
        in: formData
        name: coin_amount
        required: true
        type: string
      - description: Message
        in: formData
        name: message
        required: true
        type: string
      - description: Card Number
        in: formData
        name: card_number
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetUserSellBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: UserSell
      tags:
      - User
...
swagger: "2.0"